backend/requirements.txt
Flask
Flask-Cors
PyJWT
pymongo
python-dotenv
bcrypt
